Adding functionality in main, progress on LL

+++ Product/CentralAcquisitionProxy.h
@@ -10,6 +10,6 @@ bool disconnectFromCentralAcquisition();
 
 void selectExaminationType(const EXAMINATION_TYPES examination);
 
-bool getDoseDataFromCentralAcquisition(uint32_t * doseData);
+bool getDoseDataFromCentralAcquisition(uint16_t * doseData);
 
 #endif
